add context detail adn refactor details table
