+++ src/api/workflow.ts
@@ -109,6 +109,10 @@ export function createWorkflowDeleteMutationHook<K extends keyof paths>(path: K)
 
 export type WorkflowModel = components['schemas']['Workflow'];
 export type WorkflowRunModel = components['schemas']['WorkflowRunDetail'];
+export type AnalysisRunModel = components['schemas']['AnalysisRunDetail'];
+export type AnalysisModel = components['schemas']['Analysis'];
+export type ComputeContextModel = components['schemas']['AnalysisContext'];
+export type StorageContextModel = components['schemas']['AnalysisContext'];
 
 export const useWorkflowModel = createWorkflowQueryHook('/api/v1/workflow/');
 export const useWorkflowRunListModel = createWorkflowQueryHook('/api/v1/workflowrun/');
